+++ packages/imperative/__tests__/config/cmd/init/init.handler.test.ts
@@ -102,6 +102,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = false;
         params.arguments.globalConfig = false;
         params.arguments.prompt = true;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -145,6 +146,49 @@ describe("Configuration Initialization command handler", () => {
 
         // Secure value supplied during prompting should be on properties
         expect(ImperativeConfig.instance.config.properties.profiles.secured.properties.secret).toEqual("fakeValue");
+    });
+
+    it("should attempt to do a dry run of initializing the project configuration", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = false;
+        params.arguments.globalConfig = false;
+        params.arguments.prompt = true;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "fakeValue");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(ensureCredMgrSpy).toHaveBeenCalledTimes(1);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(1);
+        // Prompting for secure property
+        expect(promptWithTimeoutSpy).toHaveBeenCalledWith(expect.stringContaining("blank to skip:"), {"hideText": true});
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the project configuration", async () => {
@@ -212,6 +256,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = true;
         params.arguments.globalConfig = false;
         params.arguments.prompt = true;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -250,6 +295,47 @@ describe("Configuration Initialization command handler", () => {
         // Schema
         expect(writeFileSyncSpy).toHaveBeenNthCalledWith(1, fakeSchemaPath, JSON.stringify(expectedSchemaObjectNoBase, null, ConfigConstants.INDENT));
         expect(writeFileSyncSpy).toHaveBeenNthCalledWith(2, fakeProjUserPath, JSON.stringify(compObj, null, ConfigConstants.INDENT)); // Config
+    });
+
+    it("should attempt to do a dry run of initializing the project user configuration", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = true;
+        params.arguments.globalConfig = false;
+        params.arguments.prompt = true;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "fakeValue");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(ensureCredMgrSpy).toHaveBeenCalledTimes(1);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(0); // User config is a skeleton - no prompting should occur
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the project user configuration", async () => {
@@ -312,6 +398,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = false;
         params.arguments.globalConfig = true;
         params.arguments.prompt = true;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -357,6 +444,49 @@ describe("Configuration Initialization command handler", () => {
 
         // Secure value supplied during prompting should be on properties
         expect(ImperativeConfig.instance.config.properties.profiles.secured.properties.secret).toEqual("fakeValue");
+    });
+
+    it("should attempt to do a dry run of initializing the global project configuration", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = false;
+        params.arguments.globalConfig = true;
+        params.arguments.prompt = true;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "fakeValue");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(ensureCredMgrSpy).toHaveBeenCalledTimes(1);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(1);
+        // Prompting for secure property
+        expect(promptWithTimeoutSpy).toHaveBeenCalledWith(expect.stringContaining("blank to skip:"), {"hideText": true});
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the global project configuration", async () => {
@@ -426,6 +556,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = true;
         params.arguments.globalConfig = true;
         params.arguments.prompt = true;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -466,6 +597,47 @@ describe("Configuration Initialization command handler", () => {
             JSON.stringify(expectedSchemaObjectNoBase, null, ConfigConstants.INDENT)
         );
         expect(writeFileSyncSpy).toHaveBeenNthCalledWith(2, fakeGblProjUserPath, JSON.stringify(compObj, null, ConfigConstants.INDENT)); // Config
+    });
+
+    it("should attempt to do a dry run of initializing the global project user configuration", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = true;
+        params.arguments.globalConfig = true;
+        params.arguments.prompt = true;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        const ensureCredMgrSpy = jest.spyOn(handler as any, "ensureCredentialManagerLoaded");
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "fakeValue");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(ensureCredMgrSpy).toHaveBeenCalledTimes(1);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(0); // User config is a skeleton - no prompting should occur
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the global project user configuration", async () => {
@@ -530,6 +702,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = false;
         params.arguments.globalConfig = false;
         params.arguments.prompt = false;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -566,6 +739,44 @@ describe("Configuration Initialization command handler", () => {
         // Schema
         expect(writeFileSyncSpy).toHaveBeenNthCalledWith(1, fakeSchemaPath, JSON.stringify(expectedSchemaObjectNoBase, null, ConfigConstants.INDENT));
         expect(writeFileSyncSpy).toHaveBeenNthCalledWith(2, fakeProjPath, JSON.stringify(compObj, null, ConfigConstants.INDENT)); // Config
+    });
+
+    it("should attempt to do a dry run of initializing the project configuration with prompt flag false", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = false;
+        params.arguments.globalConfig = false;
+        params.arguments.prompt = false;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "fakeValue");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(0); // CI flag should not prompt
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the project configuration with prompt flag false", async () => {
@@ -626,6 +837,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = true;
         params.arguments.globalConfig = false;
         params.arguments.prompt = false;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -662,6 +874,44 @@ describe("Configuration Initialization command handler", () => {
         // Schema
         expect(writeFileSyncSpy).toHaveBeenNthCalledWith(1, fakeSchemaPath, JSON.stringify(expectedSchemaObjectNoBase, null, ConfigConstants.INDENT));
         expect(writeFileSyncSpy).toHaveBeenNthCalledWith(2, fakeProjUserPath, JSON.stringify(compObj, null, ConfigConstants.INDENT)); // Config
+    });
+
+    it("should attempt to do a dry run of initializing the project user configuration with prompting disabled", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = true;
+        params.arguments.globalConfig = false;
+        params.arguments.prompt = false;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "fakeValue");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(0); // CI flag should not prompt
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the project user configuration with prompting disabled", async () => {
@@ -722,6 +972,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = false;
         params.arguments.globalConfig = true;
         params.arguments.prompt = false;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -760,6 +1011,44 @@ describe("Configuration Initialization command handler", () => {
             JSON.stringify(expectedSchemaObjectNoBase, null, ConfigConstants.INDENT)
         );
         expect(writeFileSyncSpy).toHaveBeenNthCalledWith(2, fakeGblProjPath, JSON.stringify(compObj, null, ConfigConstants.INDENT)); // Config
+    });
+
+    it("should attempt to do a dry run of initializing the global project configuration with prompt flag false", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = false;
+        params.arguments.globalConfig = true;
+        params.arguments.prompt = false;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "fakeValue");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(0); // CI flag should not prompt
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the global project configuration with prompt flag false", async () => {
@@ -822,6 +1111,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = true;
         params.arguments.globalConfig = true;
         params.arguments.prompt = false;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -862,6 +1152,44 @@ describe("Configuration Initialization command handler", () => {
         expect(writeFileSyncSpy).toHaveBeenNthCalledWith(2, fakeGblProjUserPath,
             JSON.stringify(compObj, null, ConfigConstants.INDENT)
         ); // Config
+    });
+
+    it("should attempt to do a dry run of initializing the global project user configuration with prompting disabled", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = true;
+        params.arguments.globalConfig = true;
+        params.arguments.prompt = false;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "fakeValue");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(0); // CI flag should not prompt
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the global project user configuration with prompting disabled", async () => {
@@ -926,6 +1254,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = false;
         params.arguments.globalConfig = false;
         params.arguments.prompt = true;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -967,6 +1296,46 @@ describe("Configuration Initialization command handler", () => {
 
         // Secure value supplied during prompting should be on properties
         expect(ImperativeConfig.instance.config.properties.profiles.secured.properties.secret).toEqual(true);
+    });
+
+    it("should attempt to do a dry run of initializing the project configuration and use boolean true for the prompt", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = false;
+        params.arguments.globalConfig = false;
+        params.arguments.prompt = true;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "true");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(1);
+        // Prompting for secure property
+        expect(promptWithTimeoutSpy).toHaveBeenCalledWith(expect.stringContaining("blank to skip:"), {"hideText": true});
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the project configuration and use boolean true for the prompt", async () => {
@@ -1032,6 +1401,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = false;
         params.arguments.globalConfig = false;
         params.arguments.prompt = true;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -1073,6 +1443,46 @@ describe("Configuration Initialization command handler", () => {
 
         // Secure value supplied during prompting should be on properties
         expect(ImperativeConfig.instance.config.properties.profiles.secured.properties.secret).toEqual(false);
+    });
+
+    it("should attempt to do a dry run of initializing the project configuration and use boolean false for the prompt", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = false;
+        params.arguments.globalConfig = false;
+        params.arguments.prompt = true;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => "false");
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(1);
+        // Prompting for secure property
+        expect(promptWithTimeoutSpy).toHaveBeenCalledWith(expect.stringContaining("blank to skip:"), {"hideText": true});
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the project configuration and use boolean false for the prompt", async () => {
@@ -1138,6 +1548,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = false;
         params.arguments.globalConfig = false;
         params.arguments.prompt = true;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -1181,6 +1592,48 @@ describe("Configuration Initialization command handler", () => {
 
         // Secure value supplied during prompting should be on properties
         expect(ImperativeConfig.instance.config.properties.profiles.secured.properties.secret).toEqual(randomValueNumber);
+    });
+
+    it("should attempt to do a dry run of initializing the project configuration and use a number for the prompt", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = false;
+        params.arguments.globalConfig = false;
+        params.arguments.prompt = true;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const randomValueString = "9001";
+        const randomValueNumber = parseInt(randomValueString, 10);
+        const promptWithTimeoutSpy = jest.fn(() => randomValueString);
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(1);
+        // Prompting for secure property
+        expect(promptWithTimeoutSpy).toHaveBeenCalledWith(expect.stringContaining("blank to skip:"), {"hideText": true});
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the project configuration and use a number for the prompt", async () => {
@@ -1248,6 +1701,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = false;
         params.arguments.globalConfig = false;
         params.arguments.prompt = true;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
@@ -1289,6 +1743,46 @@ describe("Configuration Initialization command handler", () => {
 
         // Secure value supplied during prompting should be on properties
         expect(ImperativeConfig.instance.config.properties.profiles.secured.properties.secret).toEqual(undefined);
+    });
+
+    it("should attempt to do a dry run of initializing the project configuration and handle getting nothing from the prompt", async () => {
+        const handler = new InitHandler();
+        const params = getIHandlerParametersObject();
+        params.arguments.userConfig = false;
+        params.arguments.globalConfig = false;
+        params.arguments.prompt = true;
+        params.arguments.dryRun = true;
+
+        existsSyncSpy.mockReturnValue(false); // No files exist
+        searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return
+        await setupConfigToLoad(); // Setup the config
+
+        setSchemaSpy = jest.spyOn(ImperativeConfig.instance.config, "setSchema");
+
+        // We aren't testing the config initialization - clear the spies
+        existsSyncSpy.mockClear();
+        searchSpy.mockClear();
+        osHomedirSpy.mockClear();
+        currentWorkingDirectorySpy.mockClear();
+
+        // initWithSchema
+        const promptWithTimeoutSpy = jest.fn(() => undefined);
+        (params.response.console as any).prompt = promptWithTimeoutSpy;
+        writeFileSyncSpy.mockImplementation(); // Don't actually write files
+
+        // initForDryRun
+        const initForDryRunSpy = jest.spyOn(handler as any, "initForDryRun");
+
+        await handler.process(params as IHandlerParameters);
+
+        expect(promptWithTimeoutSpy).toHaveBeenCalledTimes(1);
+        // Prompting for secure property
+        expect(promptWithTimeoutSpy).toHaveBeenCalledWith(expect.stringContaining("blank to skip:"), {"hideText": true});
+
+        expect(initForDryRunSpy).toHaveBeenCalledTimes(1);
+        expect(initForDryRunSpy).toHaveBeenCalledWith(ImperativeConfig.instance.config, params.arguments.userConfig);
+
+        expect(writeFileSyncSpy).not.toHaveBeenCalled();
     });
 
     it("should attempt to overwrite the project configuration and handle getting nothing from the prompt", async () => {
@@ -1354,6 +1848,7 @@ describe("Configuration Initialization command handler", () => {
         params.arguments.userConfig = false;
         params.arguments.globalConfig = false;
         params.arguments.prompt = true;
+        params.arguments.dryRun = false;
 
         existsSyncSpy.mockReturnValue(false); // No files exist
         searchSpy.mockReturnValueOnce(fakeProjUserPath).mockReturnValueOnce(fakeProjPath); // Give search something to return